requirements.txt
aenum==2.2.6
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.28.49
botocore==1.31.49
cachetools==4.2.4
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
decorator==4.4.2
Django==4.2
django-cors-headers==3.13.0
django-filter==23.2
django-googledrive-storage==1.6.0
django-sslserver==0.22
django-storages==1.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fbchat==1.9.7
google-api-core==2.10.2
google-api-python-client==2.97.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
imageio==2.31.2
imageio-ffmpeg==0.4.8
jmespath==1.0.1
moviepy==1.0.3
multidict==6.0.4
mysqlclient==2.2.0
numpy==1.24.2
paho-mqtt==1.6.1
Pillow==10.0.0
proglog==0.1.10
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.7.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
qrcode==7.4.2
requests==2.31.0
rsa==4.9
s3transfer==0.6.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==1.4.47
sqlparse==0.4.3
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.4.0
yarl==1.8.2
